Throw an exception on a bad http return
If the token request fails with an error such as Unauthenticated or
anything else the existing code doesn't check for this code and will try
and retrieve the token anyway. This gives a somewhat difficult to debug
key error when the response doesn't contain the token data.

This is wrong, we should always be checking the response code before
trusting the response data.

This is a slight change in behaviour, we now return this exception
instead of a KeyError, however the other exception was difficult to
catch.

This reuses the failure exception from OAuth1 and makes that public.

Closes: #302
