+++ requests_oauthlib/__init__.py
@@ -1,5 +1,6 @@
 import logging
 
+from .exc import TokenRequestDenied
 from .oauth1_auth import OAuth1
 from .oauth1_session import OAuth1Session
 from .oauth2_auth import OAuth2
